fix add to cart

+++ app/Http/Controllers/CartController.php
@@ -2,28 +2,31 @@
 
 namespace App\Http\Controllers;
 
+use App\Constants\Status;
 use App\Http\Resources\CartResource;
 use App\Models\Cart;
 use App\Models\Product;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Validator;
 
 class CartController extends Controller
 {
     public function show()
     {
         $user = auth()->user();
-       return $this->success(CartResource::make($user->cart()->first()));
+        return $this->success(CartResource::make($user->cart()->first()));
     }
 
-    public function store(Product $product)
+    public function store(Product $product, Request $request)
     {
+        $validated_data = Validator::make($request->all(), ['attributes' => 'required|array',]);
+        if ($validated_data->fails())
+            return $this->error(Status::VALIDATION_FAILED, $validated_data->errors());
+
         $user = auth()->user();
-        $cart = $user->cart()->firstOrCreate([
-            'status' => 0,
-            'order_code' => 'DoNotNeed'
-        ]);
+        $cart = $user->cart()->firstOrCreate(['status' => 0]);
         $cart->products()->detach($product->id);
-        $cart->products()->attach($product->id, ['num' => '1']);
+        $cart->products()->attach($product->id, ['quantity' => '1','attributes' => json_encode($request->toArray()['attributes'])]);
         return $this->success('added to your cart');
     }
